ISIS-3207: migrate wicket viewer (model)
